backend/requirements.txt
django>=5.0
django-ninja>=1.0
django-cors-headers>=4.0
python-dotenv>=1.0
weasyprint>=60.0
pillow>=10.0
httpx>=0.27
channels>=4.0.0
chess>=1.10.0
PyMuPDF>=1.23.0
chromadb>=0.4.0
